Add files via upload
Run checker.py to train the model using the videos in training folder.
Run test.py to check the accuracy of the model. It will prompt you to enter the word that you have signed just to help with the traingin.
Run app.py to start interpreting sign language into voice.
Run one.py to run the software on the website.